Split recycling.rs (593 lines) into recycling/ subdirectory
Split crates/simulation/src/recycling.rs into focused modules under recycling/:
- tiers.rs: RecyclingTier enum and policy parameters (123 lines)
- economics.rs: RecyclingEconomics resource, constants, market cycle (101 lines)
- state.rs: RecyclingState resource (42 lines)
- systems.rs: update system and RecyclingPlugin (77 lines)
- tests.rs: all unit tests (239 lines)
- mod.rs: thin facade with re-exports (20 lines)

All public items re-exported from mod.rs — no external import changes needed.
